requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.10.19
botocore==1.13.19
branca==0.3.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
config==0.4.2
confuse==1.0.0
cycler==0.10.0
Cython==0.29.14
dataIO==0.0.5
deap==1.3.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
Flask-PyMongo==2.3.0
future==0.17.1
gast==0.2.2
gensim==3.8.1
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imbalanced-learn==0.5.0
imblearn==0.0
importlib-metadata==0.23
ipykernel==5.1.3
ipyleaflet==0.11.4
ipympl==0.3.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.2
jupyterlab-server==1.0.6
Keras==2.2.4
Keras-Applications==1.0.8
git+git://github.com/keras-team/keras-contrib.git
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langcodes==1.4.1
langdetect==1.0.7
Lasagne==0.1
lazy-property==0.0.1
lemminflect==0.2.0
librosa==0.7.1
lime==0.1.1.36
llvmlite==0.30.0
lxml==4.4.1
marisa-trie==0.7.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
MiniSom==2.2.2
mistune==0.8.4
mlxtend==0.17.0
more-itertools==7.2.0
msgpack==0.6.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.2
numpy==1.17.3
oauthlib==3.1.0
openpyxl==3.0.1
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
phe==1.4.0
pickleshare==0.7.5
Pillow==6.2.1
primesieve==1.4.4
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.5
ptyprocess==0.6.0
py-spy==0.3.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.4
pyrsistent==0.15.5
PyStemmer
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
rdflib==4.2.2
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
Send2Trash==1.5.0
seqeval==0.0.12
shap==0.32.1
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
smart-open==1.9.0
smote-variants==0.3.4
SoundFile==0.10.3.post1
statistics==1.0.3.5
stop-words==2018.7.23
stopwords==0.1.3
tabulate==0.8.6
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-addons==0.12.1
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
traittypes==0.2.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xarray==0.14.0
xgboost==0.90
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
